Remove old unused linear solvers
