+++ test/PDEs/HeatEquation.jl
@@ -1,5 +1,53 @@
 heat_eq_dir = joinpath(output_root, "HeatEquation")
 
+struct DifussionImplicit end
+function TridiagonalMatrix(
+    q::StateVec,
+    aux::StateVec,
+    Δt::T,
+    ρ::S,
+    a::S,
+    K::S,
+    grid::Grid{T},
+    ::DifussionImplicit,
+    i::Int,
+) where {S <: Symbol, T}
+    #! format: off
+    a_ = [   -Δt * grid.Δzi2 *      (aux[ρ, Dual(k)] .* q[a, Dual(k), i] .* aux[K, Dual(k), i])[2] for k in over_elems_real(grid)[1:(end - 1)]]
+    b_ = [1 + Δt * grid.Δzi2 * (sum((aux[ρ, Dual(k)] .* q[a, Dual(k), i] .* aux[K, Dual(k), i]))) for k in over_elems_real(grid)]
+    c_ = [   -Δt * grid.Δzi2 *      (aux[ρ, Dual(k)] .* q[a, Dual(k), i] .* aux[K, Dual(k), i])[1] for k in over_elems_real(grid)[2:end]]
+    #! format: on
+    return Tridiagonal(a_, b_, c_)
+end
+
+"""
+    solve_tdma!(q::StateVec, tendencies::StateVec, aux::StateVec, x::S, ρ::S, a_τp1::S, a_τ::S, K::S, grid::Grid{T}, Δt::T, i::Int=1)
+
+Solve tri-diagonal system using Diffusion-implicit time-marching.
+"""
+function solve_tdma!(
+    q::StateVec,
+    tendencies::StateVec,
+    aux::StateVec,
+    x::S,
+    ρ::S,
+    a_τp1::S,
+    a_τ::S,
+    K::S,
+    grid::Grid{T},
+    Δt::T,
+    i::Int = 1,
+) where {S <: Symbol, T}
+    B = [
+        aux[a_τ, k, i] / q[a_τp1, k, i] * q[x, k, i] +
+        Δt * tendencies[x, k, i] / (aux[ρ, k] * q[a_τp1, k, i])
+        for k in over_elems_real(grid)
+    ]
+    A = TridiagonalMatrix(q, aux, Δt, ρ, a_τp1, K, grid, DifussionImplicit(), i)
+    X = inv(A) * B
+    assign_real!(q, x, grid, X, i)
+end
+
 @testset "∂_t T = K ΔT + 1, T = 0 ∈ ∂Ω, diffusion-explicit" begin
     domain_set = DomainSet(gm = 1)
     domain_subset = DomainSubSet(gm = true)
