+++ test/runtests.jl
@@ -2,7 +2,7 @@ using Test
 
 # Unit tests / verification
 for submodule in
-    ["Grids", "DomainDecomposition", "StateVecs", "LinearSolvers", "PDEs"]
+    ["Grids", "DomainDecomposition", "StateVecs", "PDEs"]
     println("Testing $submodule")
     include(joinpath(submodule, "runtests.jl"))
 end
